chahua: orchestrator 按子域拆 5 个 slot —— 主类回归核心
仿 server.py / admin.py 的「组合而非多继承」slot 模式，把
orchestrator.py 按子域拆出 `_orchestrator_{chain,handoff_drain,
handoff_queue,managed_session,scoring,consts}.py`。主类 Orchestrator
经 `_install_orchestrator_slots(orch)` 单点装配，公开 import 路径
（`from chahua.orchestrator import Orchestrator / OrchestratorConfig`）
不变；外部直访经 property + 同名薄转发覆盖（含 30+ 测试 `orch._handoff_queue`
/ `_managed_session` / `Orchestrator._handoff_cost(...)` 等）。

主类保留 `_run_ai_chain` / `_intercept_task_proposal` 同名 method
作为绑定点：前者被 tests 经 monkeypatch.setattr 替换，后者经
session.py:511 guest.set_task_proposal_hook(...) 取 bound method 引用。

CLAUDE.md 同步在「架构要点」「handoff」「MTS」三节注 slot 归属。
